Same restructuring on chapter 3 and 4 as on 1 and 2
