+++ PlaygroundBook/Chapters/Chapter3.playgroundchapter/Pages/Page1.playgroundpage/Contents.swift
@@ -20,5 +20,5 @@ func start() {
 //#-end-hidden-code
 
 //#-hidden-code
-playgroundEpilogue( assessment() )
+playgroundEpilogue( BookProgramCommandStart.assessment )
 //#-end-hidden-code
