+++ PlaygroundBook/Chapters/Chapter3.playgroundchapter/Pages/Page2.playgroundpage/Contents.swift
@@ -25,5 +25,5 @@ func start() {
 //#-end-hidden-code
 
 //#-hidden-code
-playgroundEpilogue( assessment() )
+playgroundEpilogue( BookProgramCommandSleep.assessment )
 //#-end-hidden-code
