+++ PlaygroundBook/Chapters/Chapter3.playgroundchapter/Pages/Page7.playgroundpage/Contents.swift
@@ -21,5 +21,5 @@ func onShake() {
 //#-end-hidden-code
 
 //#-hidden-code
-playgroundEpilogue( assessment() )
+playgroundEpilogue( BookProgramCommandRandom.assessment )
 //#-end-hidden-code
