+++ PlaygroundBook/Chapters/Chapter4.playgroundchapter/Pages/Page4.playgroundpage/Contents.swift
@@ -38,5 +38,5 @@ func forever() {
 //#-end-hidden-code
 
 //#-hidden-code
-playgroundEpilogue( assessment() )
+playgroundEpilogue( BookProgramProjectClap.assessment )
 //#-end-hidden-code
