+++ PlaygroundBook/Chapters/Chapter4.playgroundchapter/Pages/Page5.playgroundpage/Contents.swift
@@ -23,5 +23,5 @@ func forever() {
 //#-end-hidden-code
 
 //#-hidden-code
-playgroundEpilogue( assessment() )
+playgroundEpilogue( BookProgramProjectTheremin.assessment )
 //#-end-hidden-code
