+++ PlaygroundBook/Chapters/Chapter4.playgroundchapter/Pages/Page6.playgroundpage/Contents.swift
@@ -36,5 +36,5 @@ func forever() {
 //#-end-hidden-code
 
 //#-hidden-code
-playgroundEpilogue( assessment() )
+playgroundEpilogue( BookProgramProjectThermometer.assessment )
 //#-end-hidden-code
