+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter3/BookProgramCommandClear.swift
@@ -1,6 +1,33 @@
 // OK
 public final class BookProgramCommandClear: ProgramBase, Program {
-    public var image: CalliopeImage = .smiley
+
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard let img = CalliopeImage(from: values[0]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let delay = UInt16(values[1]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramCommandClear()
+		p.image = img
+		p.delay = Int16(delay)
+
+		//return (.pass(message: success), p)
+		return (nil, p)
+	}
+
+	public var image: CalliopeImage = .smiley
     public var delay: Int16 = 1000
 
     public func build() -> ProgramBuildResult {
