+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter3/BookProgramCommandConditionals.swift
@@ -1,6 +1,44 @@
 // OK
 public final class BookProgramCommandConditionals: ProgramBase, Program {
-    public var textTrue: String = "w"
+
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard let start = Int(values[0]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let stop = Int(values[1]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let n = Int(values[2]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard values.count > 4 else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramCommandConditionals()
+		p.start =  Int16(start)
+		p.stop =  Int16(stop)
+		p.n =  Int16(n)
+		p.textTrue = values[3]
+		p.textFalse = values[4]
+
+		//return (.pass(message: success), p)
+		return (nil, p)
+	}
+
+	public var textTrue: String = "w"
     public var textFalse: String = "n"
     public var start: Int16 = 0
     public var stop: Int16 = 2
