+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter3/BookProgramCommandForever.swift
@@ -2,7 +2,34 @@
 import UIKit
 
 public final class BookProgramCommandForever: ProgramBase, Program {
-    public var color: UIColor = .red
+
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard let mColor = miniColor(from: values[0]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let delay = UInt16(values[1]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramCommandForever()
+		p.color = mColor.color
+		p.delay = Int16(delay)
+
+		//return (.pass(message: success), p)
+		return (nil, p)
+	}
+
+	public var color: UIColor = .red
     public var delay: Int16 = 200
 
     public func build() -> ProgramBuildResult {
