+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter3/BookProgramCommandLoops.swift
@@ -1,6 +1,38 @@
 // OK
 public final class BookProgramCommandLoops: ProgramBase, Program {
-    public var start: Int16 = 1
+
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard let start = Int(values[0]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let stop = Int(values[1]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let delay = UInt16(values[2]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramCommandLoops()
+		p.start = Int16(start)
+		p.stop = Int16(stop)
+		p.delay = Int16(delay)
+
+		//return (.pass(message: success), p)
+		return (nil, p)
+	}
+
+	public var start: Int16 = 1
     public var stop: Int16 = 3
     public var delay: Int16 = 1000
