+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter3/BookProgramCommandRandom.swift
@@ -1,6 +1,33 @@
 // OK
 public final class BookProgramCommandRandom: ProgramBase, Program {
-    public var start: Int16 = 1
+
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard let start = Int(values[0]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let stop = Int(values[1]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramCommandRandom()
+		p.start = Int16(start)
+		p.stop = Int16(stop)
+
+		//return (.pass(message: success), p)
+		return (nil, p)
+	}
+
+	public var start: Int16 = 1
     public var stop: Int16 = 6
 
     public func build() -> ProgramBuildResult {
