+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter3/BookProgramCommandSleep.swift
@@ -1,6 +1,43 @@
 // OK
 public final class BookProgramCommandSleep: ProgramBase, Program {
-    public var image1: CalliopeImage = .sad
+
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard let img1 = CalliopeImage(from: values[0]) else {
+			LogNotify.log("img1")
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let delay = UInt16(values[1]) else {
+			LogNotify.log("delay")
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		LogNotify.log("img2 \(values[2])")
+
+		guard let img2 = CalliopeImage(from: values[2]) else {
+			LogNotify.log("img2")
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramCommandSleep()
+		p.image1 = img1
+		p.delay = Int16(delay)
+		p.image2 = img2
+
+		//return (.pass(message: success), p)
+		return (nil, p)
+	}
+
+	public var image1: CalliopeImage = .sad
     public var image2: CalliopeImage = .smiley
     public var delay: Int16 = 1000
