+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter3/BookProgramCommandStart.swift
@@ -1,6 +1,28 @@
 // OK
 public final class BookProgramCommandStart: ProgramBase, Program {
-    public var s: String = "foo"
+
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard values.count > 0 else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramCommandStart()
+		p.s = values[0]
+
+		//return (.pass(message: success), p)
+		return (nil, p)
+	}
+
+	public var s: String = "foo"
 
 
     public func build() -> ProgramBuildResult {
