+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter3/BookProgramCommandVariables.swift
@@ -1,6 +1,33 @@
 // OK
 public final class BookProgramCommandVariables: ProgramBase, Program {
-    public var name: String = "Jon"
+
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard values.count > 0 else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let age = Int16(values[1]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramCommandVariables()
+		p.name = values[0]
+		p.age = age
+
+		//return (.pass(message: success), p)
+		return (nil, p)
+	}
+
+	public var name: String = "Jon"
     public var age: Int16 = 25
 
     public func build() -> ProgramBuildResult {
