+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter4/BookProgramProjectClap.swift
@@ -3,6 +3,30 @@ import UIKit
 
 public final class BookProgramProjectClap: ProgramBase, Program {
 
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard let threshold = UInt16(values[0]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+		guard let mColor = miniColor(from: values[1]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+		let p = BookProgramProjectClap()
+		p.color = mColor.color
+		p.threshold = Int16(threshold)
+
+		// return (.pass(message: success), p)
+		return (nil, p)
+	}
+
     public var color: UIColor = .white
     public var threshold: Int16 = 2
     public var delay: Int16 = 10
