+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter4/BookProgramProjectDice.swift
@@ -1,6 +1,42 @@
 // OK
 public final class BookProgramProjectDice: ProgramBase, Program {
 
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard let start = Int(values[0]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let stop = Int(values[1]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let freq = miniSound(from: values[2]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let delay = Int(values[3]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramProjectDice()
+		p.start = Int16(start)
+		p.stop = Int16(stop)
+		p.frequency = Int16(freq.rawValue)
+		p.delay = Int16(delay)
+
+		// return (.pass(message: success), p)
+		return (nil, p)
+	}
+
     public var start: Int16 = 1
     public var stop: Int16 = 6
