+++ PlaygroundBook/Sources/Playground/BookPrograms/Chapter4/BookProgramProjectTheremin.swift
@@ -1,6 +1,32 @@
 // OK
 public final class BookProgramProjectTheremin: ProgramBase, Program {
 
+	public static let assessment: AssessmentBlock = { values in
+
+		let success = "page.success".localized
+		let hints = [
+			"page.hint1".localized,
+			"page.hint2".localized,
+			"page.hint3".localized
+		]
+		let solution = "page.solution".localized
+
+		guard let factor = UInt16(values[0]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		guard let offset = UInt16(values[1]) else {
+			return (.fail(hints: hints, solution: solution), nil)
+		}
+
+		let p = BookProgramProjectTheremin()
+		p.factor = Int16(factor)
+		p.offset = Int16(offset)
+
+		// return (.pass(message: success), p)
+		return (nil, p)
+	}
+
     public var factor: Int16 = 10
     public var offset: Int16 = 200
     public var delay: Int16 = 100
